refactor: reorganiza pacotes para melhor modularidade
- Move pacotes como `properties` e `dto` para novos módulos apropriados (`api` e `internal`).
- Reorganiza implementações e interfaces para reforçar separação de responsabilidades (`ports` para `api` e `internal`).

Refs: GH-1
