+++ src/main/java/dev/ghonda/project/management/security/api/usecases/dto/LoginDetailPayload.java
@@ -1,0 +1,18 @@
+package dev.ghonda.project.management.security.api.usecases.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Builder;
+
+import java.time.LocalDateTime;
+
+@Builder
+public record LoginDetailPayload(
+    String token,
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
+    LocalDateTime expiresIn,
+    String refreshToken,
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
+    LocalDateTime refreshTokenExpiresIn,
+    String username
+) {
+}
