+++ src/main/java/dev/ghonda/project/management/security/api/usecases/dto/LoginPayload.java
@@ -1,0 +1,11 @@
+package dev.ghonda.project.management.security.api.usecases.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginPayload(
+    @NotBlank
+    String username,
+    @NotBlank
+    String password
+) {
+}
