--- src/main/java/dev/ghonda/project/management/security/constants/JwtConstants.java
@@ -1,11 +1,0 @@
-package dev.ghonda.project.management.security.constants;
-
-public final class JwtConstants {
-
-  public static final String BEARER = "Bearer ";
-
-  public static final String AUTHENTICATION_HEADER = "Authorization";
-
-  private JwtConstants() { }
-
-}
