+++ src/main/java/dev/ghonda/project/management/security/internal/constants/JwtConstants.java
@@ -1,0 +1,11 @@
+package dev.ghonda.project.management.security.internal.constants;
+
+public final class JwtConstants {
+
+  public static final String BEARER = "Bearer ";
+
+  public static final String AUTHENTICATION_HEADER = "Authorization";
+
+  private JwtConstants() { }
+
+}
