--- src/main/java/dev/ghonda/project/management/security/ports/api/services/TextEncoderService.java
@@ -1,8 +1,0 @@
-package dev.ghonda.project.management.security.ports.api.services;
-
-@FunctionalInterface
-public interface TextEncoderService {
-
-    String encode(String text);
-
-}
