--- src/main/java/dev/ghonda/project/management/security/ports/spi/service/UserDetailsServiceImpl.java
@@ -1,27 +1,0 @@
-package dev.ghonda.project.management.security.ports.spi.service;
-
-import dev.ghonda.project.management.users.domain.UserRepository;
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@Service
-@AllArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(final String username) throws UsernameNotFoundException {
-        if (log.isDebugEnabled()) { log.debug("m=loadUserByUsername(username={})", username); }
-
-        return this.userRepository.findUserByUsername(username)
-            .map(UserSecurityDetails::new)
-            .orElseThrow(() -> new UsernameNotFoundException("Usuário %s não encontrado".formatted(username)));
-    }
-
-}
