--- src/main/java/dev/ghonda/project/management/security/properties/JwtProperties.java
@@ -1,14 +1,0 @@
-package dev.ghonda.project.management.security.properties;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import java.time.Duration;
-
-@ConfigurationProperties(prefix = "security.jwt")
-public record JwtProperties(
-    String secret,
-    Duration expiration,
-    String refreshSecret,
-    Duration refreshExpiration
-) {
-}
