+++ src/main/java/dev/ghonda/project/management/users/api/usecases/RegisterUserUseCase.java
@@ -1,7 +1,7 @@
 package dev.ghonda.project.management.users.api.usecases;
 
 import dev.ghonda.project.management.shared.dto.Resource;
-import dev.ghonda.project.management.users.internal.usecases.dto.RegisterUserPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.RegisterUserPayload;
 
 @FunctionalInterface
 public interface RegisterUserUseCase {
