+++ src/main/java/dev/ghonda/project/management/users/api/usecases/SearchUserByIdUseCase.java
@@ -1,6 +1,6 @@
 package dev.ghonda.project.management.users.api.usecases;
 
-import dev.ghonda.project.management.users.internal.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
 
 @FunctionalInterface
 public interface SearchUserByIdUseCase {
