+++ src/main/java/dev/ghonda/project/management/users/api/usecases/UpdateUserUseCase.java
@@ -1,7 +1,7 @@
 package dev.ghonda.project.management.users.api.usecases;
 
-import dev.ghonda.project.management.users.internal.usecases.dto.UpdateUserPayload;
-import dev.ghonda.project.management.users.internal.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UpdateUserPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
 
 @FunctionalInterface
 public interface UpdateUserUseCase {
