+++ src/main/java/dev/ghonda/project/management/users/internal/rest/UserController.java
@@ -3,9 +3,9 @@ package dev.ghonda.project.management.users.internal.rest;
 import dev.ghonda.project.management.shared.dto.ApiCollectionPageResponse;
 import dev.ghonda.project.management.shared.dto.ApiResponse;
 import dev.ghonda.project.management.shared.dto.Resource;
-import dev.ghonda.project.management.users.internal.usecases.dto.RegisterUserPayload;
-import dev.ghonda.project.management.users.internal.usecases.dto.UpdateUserPayload;
-import dev.ghonda.project.management.users.internal.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.RegisterUserPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UpdateUserPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
 import dev.ghonda.project.management.users.api.usecases.DeleteUserUseCase;
 import dev.ghonda.project.management.users.api.usecases.RegisterUserUseCase;
 import dev.ghonda.project.management.users.api.usecases.SearchUserByIdUseCase;
