+++ src/main/java/dev/ghonda/project/management/users/internal/usecases/SearchUsersUseCaseImpl.java
@@ -2,7 +2,7 @@ package dev.ghonda.project.management.users.internal.usecases;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
 import dev.ghonda.project.management.users.domain.UserRepository;
-import dev.ghonda.project.management.users.internal.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
 import dev.ghonda.project.management.users.api.usecases.SearchUsersUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
