+++ src/main/java/dev/ghonda/project/management/users/internal/usecases/UpdateUserUseCaseImpl.java
@@ -5,8 +5,8 @@ import dev.ghonda.project.management.shared.exceptions.ResourceNotFoundException
 import dev.ghonda.project.management.shared.validators.ValidatorService;
 import dev.ghonda.project.management.users.domain.Role;
 import dev.ghonda.project.management.users.domain.UserRepository;
-import dev.ghonda.project.management.users.internal.usecases.dto.UpdateUserPayload;
-import dev.ghonda.project.management.users.internal.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UpdateUserPayload;
+import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
 import dev.ghonda.project.management.users.api.usecases.UpdateUserUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
