--- src/main/java/dev/ghonda/project/management/users/internal/usecases/dto/UpdateUserPayload.java
@@ -1,13 +1,0 @@
-package dev.ghonda.project.management.users.internal.usecases.dto;
-
-import dev.ghonda.project.management.shared.validators.annotations.EnumValue;
-import dev.ghonda.project.management.users.domain.Role;
-
-public record UpdateUserPayload(
-    String username,
-    String fullName,
-    String email,
-    @EnumValue(enumClass = Role.class)
-    String role
-) {
-}
